Refactor CSV download functions to reduce code duplication
- Create csvDownloadCommon shared function with configurable options
- Refactor csvDownload to use common function with default options
- Refactor csvDownloadForWindows to use common function with BOM and Excel suffix
- Improve code readability with early return pattern
- Reduce duplicate code from ~50 lines to ~15 lines per function
